Address Simon's PR feedback - add new source file TestExclusion.cs
I concur with Simon that attaching the new test exclusion functionality
to the existing SuperIlcHelpers was somewhat hacky so I decoupled
it from the BuildOptions class and put it into a new file.

Thanks

Tomas
